refactor(billing): isolate simulator calculation loading
